first implementation of most important connections, bugfixes
- I implemented the most important connections between Node and NodeItem, so they don't have to methods on each other anymore
- introduced NodeAction-trigger-system such that node actions automatically switch from the GUI thread to the flow worker thread

Some record:
I just did some rudimentary tests, adapted a few nodes to the system, including the OpenCV package. There are still a few smaller bugs in the core, but the threading works in principal, if the widgets are implemented correctly. There are issues with the synchronization of the flows (animations don't execute correctly anymore etc.) which might be harder to fix, but it's probably fixable.
I don't know if I like this system. It is incredibly powerful, but programming widgets can get really hard this way. I will make a few more tests, try to sort out the GUI issues and hopefully see if the threading idea is worth the struggle.