flow splitting halfway done
The Flow class itself is now mostly split. I already implementet the necessary join mechanisms for creating config_data in the script as well as in the FlowWidget. Coming up is the reimplementation of undo/redo operations. It will show whether the split works nicely with the current threading approach. If this is successful, I could finish the threading prototype soon.