raise version
So far, the version minor and micro were analogous to Ryven's major and minor respectively for compatible versions. Due to the committed and incoming large changes only in ryvencore and further decoupled development, I am breaking this consistency for now. I might raise Ryven's version later to rematch if appropriate.