DType add-on prototype
This adds a prototype implementation for a DType add-on. However, it does absolutely nothing in action. So far, DTypes were motivated by adding convenient standard widgets in ryvencore-qt. Because without certain related extensions of the Node class this add-on doesn't make much sense, I will move it to ryvencore-qt or Ryven.

For reference, I am comitting this here first.